refactor(application): GameController 생성 및 기능 분리
